+++ dkutils/datakitchen_api/recipe.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import logging
 
+from requests import Response
 from typing import TYPE_CHECKING
 from pathlib import Path
 
@@ -52,6 +53,14 @@ class Recipe:
         -------
         Recipe
             :class:`Recipe <Recipe>` object
+
+        Raises
+        ------
+        HTTPError
+            If the request fails.
+        ValueError
+            If recipe_name is empty or None
+            If recipe_name already exists in the provided kitchen
         """
         logger.debug(f'Creating recipe named {recipe_name} in kitchen {client.kitchen}...')
         client._ensure_attributes(KITCHEN)
@@ -69,7 +78,7 @@ class Recipe:
         )
         return Recipe(client, recipe_name)
 
-    def delete(self, kitchen_name: str):
+    def delete(self, kitchen_name: str) -> Response:
         """
         Delete this recipe from the provided kitchen.
 
@@ -82,12 +91,16 @@ class Recipe:
         -------
         requests.Response
             :class:`Response <Response>` object
-        """
-        logger.debug(f'Deleting recipe named {self._name} in kitchen {kitchen_name}...')
-        response = self._client._api_request(API_DELETE, 'recipe', kitchen_name, self._name)
-        return response
 
-    def get_recipe_files(self, kitchen_name: str):
+        Raises
+        ------
+        HTTPError
+            If the request fails.
+       """
+        logger.debug(f'Deleting recipe named {self._name} in kitchen {kitchen_name}...')
+        return self._client._api_request(API_DELETE, 'recipe', kitchen_name, self._name)
+
+    def get_recipe_files(self, kitchen_name: str) -> dict:
         """
         Retrieve all the files for this recipe in the provided kitchen.
 
@@ -100,6 +113,13 @@ class Recipe:
         -------
         dict
             Dictionary keyed by file path and valued by file contents string.
+
+        Raises
+        ------
+        HTTPError
+            If the request fails.
+        Exception
+            If a filetype is unrecognized.
         """
         logger.debug(f'Retrieving files for recipe {self._name} in kitchen {kitchen_name}...')
         response = self._client._api_request(API_GET, 'recipe', 'get', kitchen_name, self._name)
@@ -123,7 +143,7 @@ class Recipe:
 
         return recipe_files_dict
 
-    def update_recipe_files(self, kitchen_name: str, filepaths: dict):
+    def update_recipe_files(self, kitchen_name: str, filepaths: dict) -> Response:
         """
         Update the files for this recipe in the provided kitchen.
 
@@ -138,6 +158,11 @@ class Recipe:
         -------
         requests.Response
             :class:`Response <Response>` object
+
+        Raises
+        ------
+        HTTPError
+            If the request fails.
         """
         logger.debug(
             f'Updating files ({list(filepaths.keys())}) for recipe {self._name} in kitchen {kitchen_name}...'
@@ -150,7 +175,7 @@ class Recipe:
         for p, c in filepaths.items():
             files[p] = {'contents': c, 'isNew': False if p in recipe_files else True}
 
-        response = self._client._api_request(
+        return self._client._api_request(
             API_POST,
             'recipe',
             'update',
@@ -161,9 +186,8 @@ class Recipe:
             files=files,
             message=f'Creating recipe files {files.keys()}'
         )
-        return response
 
-    def delete_recipe_files(self, kitchen_name: str, filepaths: list):
+    def delete_recipe_files(self, kitchen_name: str, filepaths: list) -> Response:
         """
         Delete the provided files from this recipe in the provided kitchen.
 
@@ -178,6 +202,11 @@ class Recipe:
         -------
         requests.Response
             :class:`Response <Response>` object
+
+        Raises
+        ------
+        HTTPError
+            If the request fails.
         """
         logger.debug(
             f'Deleting files ({filepaths}) for recipe {self._name} in kitchen {kitchen_name}...'
@@ -186,7 +215,7 @@ class Recipe:
         # Including an empty dictionary for a file path implies file deletion
         files = {p: {} for p in filepaths}
 
-        response = self._client._api_request(
+        return self._client._api_request(
             API_POST,
             'recipe',
             'update',
@@ -197,4 +226,3 @@ class Recipe:
             files=files,
             message=f'Deleting recipe files {filepaths}'
         )
-        return response